requirements.txt
beautifulsoup4==4.12.3
pymongo==4.7.2
kafka-python==2.0.2
requests==2.31.0
schedule==1.2.2
Flask==3.1.1
httpx==0.28.1